Introduce LentCheckOut.
Given a `CheckOut` of a resource, `LentCheckOut` is an asynchronous computation
where the resource is injected, then reclaimed when it completes. Since that
computation is opaque, there is no way to reclaim the resource in the face of a
failure. `LentCheckOut` handles this by accepting the situation and decrementing
the resource creation count, allowing other computations to create new
resources.